[REFACTOR] - Update admin sub-component imports to (app) route group path
